Move built-in condition types to impl
