+++ fabric-resource-conditions-api-v1/src/main/java/net/fabricmc/fabric/api/resource/conditions/v1/ResourceConditionType.java
@@ -4,16 +4,6 @@ import com.mojang.serialization.Codec;
 
 import com.mojang.serialization.DataResult;
 
-import net.fabricmc.fabric.api.resource.conditions.v1.conditions.AllModsLoadedResourceCondition;
-import net.fabricmc.fabric.api.resource.conditions.v1.conditions.AndResourceCondition;
-import net.fabricmc.fabric.api.resource.conditions.v1.conditions.AnyModsLoadedResourceCondition;
-import net.fabricmc.fabric.api.resource.conditions.v1.conditions.FeaturesEnabledResourceCondition;
-import net.fabricmc.fabric.api.resource.conditions.v1.conditions.NotResourceCondition;
-import net.fabricmc.fabric.api.resource.conditions.v1.conditions.OrResourceCondition;
-import net.fabricmc.fabric.api.resource.conditions.v1.conditions.RegistryContainsResourceCondition;
-import net.fabricmc.fabric.api.resource.conditions.v1.conditions.TagsPopulatedResourceCondition;
-import net.fabricmc.fabric.api.resource.conditions.v1.conditions.TrueResourceCondition;
-
 import net.minecraft.util.Identifier;
 
 import java.util.Optional;
