+++ fabric-resource-conditions-api-v1/src/main/java/net/fabricmc/fabric/impl/resource/conditions/DefaultResourceConditionTypes.java
@@ -1,16 +1,19 @@
-package net.fabricmc.fabric.api.resource.conditions.v1;
+package net.fabricmc.fabric.impl.resource.conditions;
 
 import com.mojang.serialization.Codec;
 
-import net.fabricmc.fabric.api.resource.conditions.v1.conditions.AllModsLoadedResourceCondition;
-import net.fabricmc.fabric.api.resource.conditions.v1.conditions.AndResourceCondition;
-import net.fabricmc.fabric.api.resource.conditions.v1.conditions.AnyModsLoadedResourceCondition;
-import net.fabricmc.fabric.api.resource.conditions.v1.conditions.FeaturesEnabledResourceCondition;
-import net.fabricmc.fabric.api.resource.conditions.v1.conditions.NotResourceCondition;
-import net.fabricmc.fabric.api.resource.conditions.v1.conditions.OrResourceCondition;
-import net.fabricmc.fabric.api.resource.conditions.v1.conditions.RegistryContainsResourceCondition;
-import net.fabricmc.fabric.api.resource.conditions.v1.conditions.TagsPopulatedResourceCondition;
-import net.fabricmc.fabric.api.resource.conditions.v1.conditions.TrueResourceCondition;
+import net.fabricmc.fabric.api.resource.conditions.v1.ResourceCondition;
+import net.fabricmc.fabric.api.resource.conditions.v1.ResourceConditionType;
+import net.fabricmc.fabric.api.resource.conditions.v1.ResourceConditions;
+import net.fabricmc.fabric.impl.resource.conditions.conditions.AllModsLoadedResourceCondition;
+import net.fabricmc.fabric.impl.resource.conditions.conditions.AndResourceCondition;
+import net.fabricmc.fabric.impl.resource.conditions.conditions.AnyModsLoadedResourceCondition;
+import net.fabricmc.fabric.impl.resource.conditions.conditions.FeaturesEnabledResourceCondition;
+import net.fabricmc.fabric.impl.resource.conditions.conditions.NotResourceCondition;
+import net.fabricmc.fabric.impl.resource.conditions.conditions.OrResourceCondition;
+import net.fabricmc.fabric.impl.resource.conditions.conditions.RegistryContainsResourceCondition;
+import net.fabricmc.fabric.impl.resource.conditions.conditions.TagsPopulatedResourceCondition;
+import net.fabricmc.fabric.impl.resource.conditions.conditions.TrueResourceCondition;
 
 import net.minecraft.util.Identifier;
