+++ fabric-resource-conditions-api-v1/src/main/java/net/fabricmc/fabric/impl/resource/conditions/conditions/AndResourceCondition.java
@@ -1,11 +1,11 @@
-package net.fabricmc.fabric.api.resource.conditions.v1.conditions;
+package net.fabricmc.fabric.impl.resource.conditions.conditions;
 
 import java.util.List;
 
 import com.mojang.serialization.Codec;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
 
-import net.fabricmc.fabric.api.resource.conditions.v1.DefaultResourceConditionTypes;
+import net.fabricmc.fabric.impl.resource.conditions.DefaultResourceConditionTypes;
 import net.fabricmc.fabric.api.resource.conditions.v1.ResourceCondition;
 import net.fabricmc.fabric.api.resource.conditions.v1.ResourceConditionType;
 import net.fabricmc.fabric.impl.resource.conditions.ResourceConditionsImpl;
